+++ forge/src/main/java/com/additional_jewelry/forge/compat/CompatFeatures.java
@@ -1,0 +1,9 @@
+package com.additional_jewelry.forge.compat;
+
+import com.additional_jewelry.forge.compat.curios.CuriosCompat;
+
+public class CompatFeatures {
+    public static void init() {
+        CuriosCompat.init();
+    }
+}